BackgroundRemover/requirements.txt
streamlit
mediapipe==0.10.14
opencv-python-headless
numpy
pillow